[DCA] Add Prometheus AD
