+++ pkg/autodiscovery/providers/names/provider_names.go
@@ -7,18 +7,19 @@ package names
 
 // User-facing names for the config providers
 const (
-	Consul          = "consul"
-	CloudFoundryBBS = "cloudfoundry-bbs"
-	ClusterChecks   = "cluster-checks"
-	Docker          = "docker"
-	ECS             = "ecs"
-	EndpointsChecks = "endpoints-checks"
-	Etcd            = "etcd"
-	File            = "file"
-	Kubernetes      = "kubernetes"
-	KubeServices    = "kubernetes-services"
-	KubeEndpoints   = "kubernetes-endpoints"
-	Prometheus      = "prometheus"
-	SNMP            = "snmp"
-	Zookeeper       = "zookeeper"
+	Consul             = "consul"
+	CloudFoundryBBS    = "cloudfoundry-bbs"
+	ClusterChecks      = "cluster-checks"
+	Docker             = "docker"
+	ECS                = "ecs"
+	EndpointsChecks    = "endpoints-checks"
+	Etcd               = "etcd"
+	File               = "file"
+	Kubernetes         = "kubernetes"
+	KubeServices       = "kubernetes-services"
+	KubeEndpoints      = "kubernetes-endpoints"
+	PrometheusPods     = "prometheus-pods"
+	PrometheusServices = "prometheus-services"
+	SNMP               = "snmp"
+	Zookeeper          = "zookeeper"
 )
