+++ pkg/autodiscovery/providers/prometheus_common_test.go
@@ -6,10 +6,209 @@
 package providers
 
 import (
+	"regexp"
 	"testing"
 
+	"github.com/DataDog/datadog-agent/pkg/config"
 	"github.com/stretchr/testify/assert"
 )
+
+func TestSetupConfigs(t *testing.T) {
+	tests := []struct {
+		name       string
+		config     []*PrometheusCheck
+		wantChecks []*PrometheusCheck
+		wantErr    bool
+	}{
+		{
+			name:   "empty config, default check",
+			config: []*PrometheusCheck{},
+			wantChecks: []*PrometheusCheck{
+				{
+					Instances: []*OpenmetricsInstance{
+						{
+							Metrics:   []string{"*"},
+							Namespace: "",
+						},
+					},
+					AD: &ADConfig{
+						ExcludeAutoconf: boolPointer(true),
+						KubeAnnotations: &InclExcl{
+							Excl: map[string]string{"prometheus.io/scrape": "false"},
+							Incl: map[string]string{"prometheus.io/scrape": "true"},
+						},
+						KubeContainerNames: []string{},
+					},
+				},
+			},
+			wantErr: false,
+		},
+		{
+			name: "custom instance, set required fields",
+			config: []*PrometheusCheck{
+				{
+					Instances: []*OpenmetricsInstance{
+						{
+							Timeout: 20,
+						},
+					},
+				},
+			},
+			wantChecks: []*PrometheusCheck{
+				{
+					Instances: []*OpenmetricsInstance{
+						{
+							Metrics:   []string{"*"},
+							Namespace: "",
+							Timeout:   20,
+						},
+					},
+					AD: &ADConfig{
+						ExcludeAutoconf: boolPointer(true),
+						KubeAnnotations: &InclExcl{
+							Excl: map[string]string{"prometheus.io/scrape": "false"},
+							Incl: map[string]string{"prometheus.io/scrape": "true"},
+						},
+						KubeContainerNames: []string{},
+					},
+				},
+			},
+			wantErr: false,
+		},
+		{
+			name: "custom AD, set required fields",
+			config: []*PrometheusCheck{
+				{
+					AD: &ADConfig{
+						KubeAnnotations: &InclExcl{
+							Excl: map[string]string{"custom/annotation": "exclude"},
+						},
+						KubeContainerNames: []string{"foo*"},
+					},
+				},
+			},
+			wantChecks: []*PrometheusCheck{
+				{
+					Instances: []*OpenmetricsInstance{
+						{
+							Metrics:   []string{"*"},
+							Namespace: "",
+						},
+					},
+					AD: &ADConfig{
+						ExcludeAutoconf: boolPointer(true),
+						KubeAnnotations: &InclExcl{
+							Excl: map[string]string{"custom/annotation": "exclude"},
+							Incl: map[string]string{"prometheus.io/scrape": "true"},
+						},
+						KubeContainerNames: []string{"foo*"},
+						containersRe:       regexp.MustCompile("foo*"),
+					},
+				},
+			},
+			wantErr: false,
+		},
+		{
+			name: "custom instances and AD",
+			config: []*PrometheusCheck{
+				{
+					Instances: []*OpenmetricsInstance{
+						{
+							Metrics:       []string{"foo", "bar"},
+							Namespace:     "custom_ns",
+							IgnoreMetrics: []string{"baz"},
+						},
+					},
+					AD: &ADConfig{
+						ExcludeAutoconf: boolPointer(false),
+						KubeAnnotations: &InclExcl{
+							Incl: map[string]string{"custom/annotation": "include"},
+							Excl: map[string]string{"custom/annotation": "exclude"},
+						},
+						KubeContainerNames: []string{},
+					},
+				},
+			},
+			wantChecks: []*PrometheusCheck{
+				{
+					Instances: []*OpenmetricsInstance{
+						{
+							Metrics:       []string{"foo", "bar"},
+							Namespace:     "custom_ns",
+							IgnoreMetrics: []string{"baz"},
+						},
+					},
+					AD: &ADConfig{
+						ExcludeAutoconf: boolPointer(false),
+						KubeAnnotations: &InclExcl{
+							Incl: map[string]string{"custom/annotation": "include"},
+							Excl: map[string]string{"custom/annotation": "exclude"},
+						},
+						KubeContainerNames: []string{},
+					},
+				},
+			},
+			wantErr: false,
+		},
+		{
+			name: "invalid check",
+			config: []*PrometheusCheck{
+				{
+					AD: &ADConfig{
+						KubeContainerNames: []string{"*"},
+					},
+				},
+			},
+			wantChecks: nil,
+			wantErr:    false,
+		},
+		{
+			name: "two checks, one invalid check",
+			config: []*PrometheusCheck{
+				{
+					AD: &ADConfig{
+						KubeContainerNames: []string{"*"},
+					},
+				},
+				{
+					AD: &ADConfig{
+						KubeContainerNames: []string{"foo", "bar"},
+					},
+				},
+			},
+			wantChecks: []*PrometheusCheck{
+				{
+					Instances: []*OpenmetricsInstance{
+						{
+							Metrics:   []string{"*"},
+							Namespace: "",
+						},
+					},
+					AD: &ADConfig{
+						ExcludeAutoconf: boolPointer(true),
+						KubeAnnotations: &InclExcl{
+							Excl: map[string]string{"prometheus.io/scrape": "false"},
+							Incl: map[string]string{"prometheus.io/scrape": "true"},
+						},
+						KubeContainerNames: []string{"foo", "bar"},
+						containersRe:       regexp.MustCompile("foo|bar"),
+					},
+				},
+			},
+			wantErr: false,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			p := &PrometheusConfigProvider{}
+			config.Datadog.Set("prometheus_scrape.checks", tt.config)
+			if err := p.setupConfigs(); (err != nil) != tt.wantErr {
+				t.Errorf("PrometheusConfigProvider.setupConfigs() error = %v, wantErr %v", err, tt.wantErr)
+			}
+			assert.EqualValues(t, tt.wantChecks, p.checks)
+		})
+	}
+}
 
 func TestBuildURL(t *testing.T) {
 	tests := []struct {
