+++ pkg/autodiscovery/providers/prometheus_pods_test.go
@@ -8,213 +8,14 @@
 package providers
 
 import (
-	"regexp"
 	"testing"
 
 	"github.com/DataDog/datadog-agent/pkg/autodiscovery/integration"
 	"github.com/DataDog/datadog-agent/pkg/autodiscovery/providers/names"
-	"github.com/DataDog/datadog-agent/pkg/config"
 	"github.com/DataDog/datadog-agent/pkg/util/kubernetes/kubelet"
 
 	"github.com/stretchr/testify/assert"
 )
-
-func TestSetupConfigs(t *testing.T) {
-	tests := []struct {
-		name       string
-		config     []*PrometheusCheck
-		wantChecks []*PrometheusCheck
-		wantErr    bool
-	}{
-		{
-			name:   "empty config, default check",
-			config: []*PrometheusCheck{},
-			wantChecks: []*PrometheusCheck{
-				{
-					Instances: []*OpenmetricsInstance{
-						{
-							Metrics:   []string{"*"},
-							Namespace: "",
-						},
-					},
-					AD: &ADConfig{
-						ExcludeAutoconf: boolPointer(true),
-						KubeAnnotations: &InclExcl{
-							Excl: map[string]string{"prometheus.io/scrape": "false"},
-							Incl: map[string]string{"prometheus.io/scrape": "true"},
-						},
-						KubeContainerNames: []string{},
-					},
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "custom instance, set required fields",
-			config: []*PrometheusCheck{
-				{
-					Instances: []*OpenmetricsInstance{
-						{
-							Timeout: 20,
-						},
-					},
-				},
-			},
-			wantChecks: []*PrometheusCheck{
-				{
-					Instances: []*OpenmetricsInstance{
-						{
-							Metrics:   []string{"*"},
-							Namespace: "",
-							Timeout:   20,
-						},
-					},
-					AD: &ADConfig{
-						ExcludeAutoconf: boolPointer(true),
-						KubeAnnotations: &InclExcl{
-							Excl: map[string]string{"prometheus.io/scrape": "false"},
-							Incl: map[string]string{"prometheus.io/scrape": "true"},
-						},
-						KubeContainerNames: []string{},
-					},
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "custom AD, set required fields",
-			config: []*PrometheusCheck{
-				{
-					AD: &ADConfig{
-						KubeAnnotations: &InclExcl{
-							Excl: map[string]string{"custom/annotation": "exclude"},
-						},
-						KubeContainerNames: []string{"foo*"},
-					},
-				},
-			},
-			wantChecks: []*PrometheusCheck{
-				{
-					Instances: []*OpenmetricsInstance{
-						{
-							Metrics:   []string{"*"},
-							Namespace: "",
-						},
-					},
-					AD: &ADConfig{
-						ExcludeAutoconf: boolPointer(true),
-						KubeAnnotations: &InclExcl{
-							Excl: map[string]string{"custom/annotation": "exclude"},
-							Incl: map[string]string{"prometheus.io/scrape": "true"},
-						},
-						KubeContainerNames: []string{"foo*"},
-						containersRe:       regexp.MustCompile("foo*"),
-					},
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "custom instances and AD",
-			config: []*PrometheusCheck{
-				{
-					Instances: []*OpenmetricsInstance{
-						{
-							Metrics:       []string{"foo", "bar"},
-							Namespace:     "custom_ns",
-							IgnoreMetrics: []string{"baz"},
-						},
-					},
-					AD: &ADConfig{
-						ExcludeAutoconf: boolPointer(false),
-						KubeAnnotations: &InclExcl{
-							Incl: map[string]string{"custom/annotation": "include"},
-							Excl: map[string]string{"custom/annotation": "exclude"},
-						},
-						KubeContainerNames: []string{},
-					},
-				},
-			},
-			wantChecks: []*PrometheusCheck{
-				{
-					Instances: []*OpenmetricsInstance{
-						{
-							Metrics:       []string{"foo", "bar"},
-							Namespace:     "custom_ns",
-							IgnoreMetrics: []string{"baz"},
-						},
-					},
-					AD: &ADConfig{
-						ExcludeAutoconf: boolPointer(false),
-						KubeAnnotations: &InclExcl{
-							Incl: map[string]string{"custom/annotation": "include"},
-							Excl: map[string]string{"custom/annotation": "exclude"},
-						},
-						KubeContainerNames: []string{},
-					},
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "invalid check",
-			config: []*PrometheusCheck{
-				{
-					AD: &ADConfig{
-						KubeContainerNames: []string{"*"},
-					},
-				},
-			},
-			wantChecks: nil,
-			wantErr:    false,
-		},
-		{
-			name: "two checks, one invalid check",
-			config: []*PrometheusCheck{
-				{
-					AD: &ADConfig{
-						KubeContainerNames: []string{"*"},
-					},
-				},
-				{
-					AD: &ADConfig{
-						KubeContainerNames: []string{"foo", "bar"},
-					},
-				},
-			},
-			wantChecks: []*PrometheusCheck{
-				{
-					Instances: []*OpenmetricsInstance{
-						{
-							Metrics:   []string{"*"},
-							Namespace: "",
-						},
-					},
-					AD: &ADConfig{
-						ExcludeAutoconf: boolPointer(true),
-						KubeAnnotations: &InclExcl{
-							Excl: map[string]string{"prometheus.io/scrape": "false"},
-							Incl: map[string]string{"prometheus.io/scrape": "true"},
-						},
-						KubeContainerNames: []string{"foo", "bar"},
-						containersRe:       regexp.MustCompile("foo|bar"),
-					},
-				},
-			},
-			wantErr: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			p := &PrometheusPodsConfigProvider{}
-			config.Datadog.Set("prometheus_scrape.checks", tt.config)
-			if err := p.setupConfigs(); (err != nil) != tt.wantErr {
-				t.Errorf("PrometheusPodsConfigProvider.setupConfigs() error = %v, wantErr %v", err, tt.wantErr)
-			}
-			assert.EqualValues(t, tt.wantChecks, p.checks)
-		})
-	}
-}
 
 func TestConfigsForPod(t *testing.T) {
 	tests := []struct {
@@ -252,7 +53,7 @@ func TestConfigsForPod(t *testing.T) {
 					Name:          "openmetrics",
 					InitConfig:    integration.Data("{}"),
 					Instances:     []integration.Data{integration.Data("{\"prometheus_url\":\"http://%%host%%:%%port%%/metrics\",\"namespace\":\"\",\"metrics\":[\"*\"]}")},
-					Provider:      names.Prometheus,
+					Provider:      names.PrometheusPods,
 					Source:        "kubelet:foo-ctr-id",
 					ADIdentifiers: []string{"foo-ctr-id"},
 				},
@@ -294,7 +95,7 @@ func TestConfigsForPod(t *testing.T) {
 					Name:          "openmetrics",
 					InitConfig:    integration.Data("{}"),
 					Instances:     []integration.Data{integration.Data("{\"prometheus_url\":\"foo/bar\",\"namespace\":\"\",\"metrics\":[\"*\"]}")},
-					Provider:      names.Prometheus,
+					Provider:      names.PrometheusPods,
 					Source:        "kubelet:foo-ctr-id",
 					ADIdentifiers: []string{"foo-ctr-id"},
 				},
@@ -386,7 +187,7 @@ func TestConfigsForPod(t *testing.T) {
 					Name:          "openmetrics",
 					InitConfig:    integration.Data("{}"),
 					Instances:     []integration.Data{integration.Data("{\"prometheus_url\":\"http://%%host%%:%%port%%/metrics\",\"namespace\":\"\",\"metrics\":[\"*\"]}")},
-					Provider:      names.Prometheus,
+					Provider:      names.PrometheusPods,
 					Source:        "kubelet:foo-ctr1-id",
 					ADIdentifiers: []string{"foo-ctr1-id"},
 				},
@@ -394,7 +195,7 @@ func TestConfigsForPod(t *testing.T) {
 					Name:          "openmetrics",
 					InitConfig:    integration.Data("{}"),
 					Instances:     []integration.Data{integration.Data("{\"prometheus_url\":\"http://%%host%%:%%port%%/metrics\",\"namespace\":\"\",\"metrics\":[\"*\"]}")},
-					Provider:      names.Prometheus,
+					Provider:      names.PrometheusPods,
 					Source:        "kubelet:foo-ctr2-id",
 					ADIdentifiers: []string{"foo-ctr2-id"},
 				},
@@ -440,7 +241,7 @@ func TestConfigsForPod(t *testing.T) {
 					Name:          "openmetrics",
 					InitConfig:    integration.Data("{}"),
 					Instances:     []integration.Data{integration.Data("{\"prometheus_url\":\"http://%%host%%:%%port%%/metrics\",\"namespace\":\"\",\"metrics\":[\"*\"]}")},
-					Provider:      names.Prometheus,
+					Provider:      names.PrometheusPods,
 					Source:        "kubelet:foo-ctr1-id",
 					ADIdentifiers: []string{"foo-ctr1-id"},
 				},
@@ -507,7 +308,7 @@ func TestConfigsForPod(t *testing.T) {
 					Name:          "openmetrics",
 					InitConfig:    integration.Data("{}"),
 					Instances:     []integration.Data{integration.Data("{\"prometheus_url\":\"http://%%host%%:%%port%%/metrics\",\"namespace\":\"\",\"metrics\":[\"*\"]}")},
-					Provider:      names.Prometheus,
+					Provider:      names.PrometheusPods,
 					Source:        "kubelet:foo-ctr-id",
 					ADIdentifiers: []string{"foo-ctr-id"},
 				},
